refactor: 설문 응답 및 응답한 결과 DTO inner class로 변경
