+++ survwey/survwey/src/main/java/mcnc/survwey/api/survey/response/service/SurveyReplyService.java
@@ -8,7 +8,6 @@ import mcnc.survwey.domain.objAnswer.repository.ObjAnswerRepository;
 import mcnc.survwey.domain.objAnswer.service.ObjAnswerService;
 import mcnc.survwey.domain.question.Question;
 import mcnc.survwey.domain.respond.Respond;
-import mcnc.survwey.api.survey.response.dto.reply.ReplyDTO;
 import mcnc.survwey.domain.respond.repository.RespondRepository;
 import mcnc.survwey.domain.respond.service.RespondService;
 import mcnc.survwey.domain.subjAnswer.SubjAnswer;
@@ -24,6 +23,7 @@ import mcnc.survwey.global.exception.custom.ErrorCode;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -67,7 +67,7 @@ public class SurveyReplyService {
         //설문이 만료되었는지 확인
         surveyService.checkSurveyExpiration(respondedSurvey.getExpireDate());
 
-        List<ReplyDTO> responseList = surveyReplyDTO.getResponseList();
+        List<SurveyReplyDTO.ReplyResponseDTO> responseList = surveyReplyDTO.getResponseList();
         //객관식, 주관식 응답 생성
         List<ObjAnswer> objAnswerList = objAnswerService.createObjectiveAnswers(responseList, respondedUser);
         List<SubjAnswer> subjAnswerList = subjAnswerService.createSubjectiveAnswers(responseList, respondedUser);
@@ -90,7 +90,7 @@ public class SurveyReplyService {
                 .stream().map(Question::getQuesId).collect(Collectors.toSet());
 
         Set<Long> existingQuestionSet = surveyReplyDTO.getResponseList()
-                .stream().map(ReplyDTO::getQuesId).collect(Collectors.toSet());
+                .stream().map(SurveyReplyDTO.ReplyResponseDTO::getQuesId).collect(Collectors.toSet());
 
         if (!inputQuestionSet.equals(existingQuestionSet)) {
             throw new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.QUESTION_NOT_MATCH_TO_SURVEY);
