+++ survwey/survwey/src/main/java/mcnc/survwey/domain/subjAnswer/service/SubjAnswerService.java
@@ -1,10 +1,10 @@
 package mcnc.survwey.domain.subjAnswer.service;
 
 import lombok.RequiredArgsConstructor;
+import mcnc.survwey.api.survey.response.dto.reply.SurveyReplyDTO;
 import mcnc.survwey.domain.question.enums.QuestionType;
 import mcnc.survwey.domain.question.Question;
 import mcnc.survwey.domain.question.service.QuestionService;
-import mcnc.survwey.api.survey.response.dto.reply.ReplyDTO;
 import mcnc.survwey.domain.subjAnswer.SubjAnswer;
 import mcnc.survwey.domain.user.User;
 import org.springframework.stereotype.Service;
@@ -26,7 +26,7 @@ public class SubjAnswerService {
      * @param respondedUser
      * @return
      */
-    public List<SubjAnswer> createSubjectiveAnswers(List<ReplyDTO> responseList, User respondedUser) {
+    public List<SubjAnswer> createSubjectiveAnswers(List<SurveyReplyDTO.ReplyResponseDTO> responseList, User respondedUser) {
         return responseList.stream()
                 .filter(replyDTO -> replyDTO.getQuestionType() == QuestionType.SUBJECTIVE)
                 .map(replyDTO -> {
